Re-enable macro creation for custom rolls.
There is an as-yet-unsolved problem in that you can create a macro from a token character sheet not linked to the actual character, meaning you can create a macro that tries to roll a custom roll that only exists on some token version of a character and not the actual character. This is a general problem and I need to figure out the most reasonable general solution.