Update verified Foundry version to 13.347.
A quick spot check doesn't show any functional breaks when using the system module. There are some deprecation warnings that will need to be addressed when v14 and v15 hit. Deferring those for now in order to delay the point at which we might need to fork to have compatible system versions for all Foundry releases from v11 forward.